+++ src/main/kotlin/io/github/jamalam360/Bot.kt
@@ -2,10 +2,7 @@ package io.github.jamalam360
 
 import com.kotlindiscord.kord.extensions.ExtensibleBot
 import io.github.jamalam360.extensions.bot.BotUtilityExtension
-import io.github.jamalam360.extensions.moderation.LoggingExtension
-import io.github.jamalam360.extensions.moderation.ModerationExtension
-import io.github.jamalam360.extensions.moderation.ModuleExtension
-import io.github.jamalam360.extensions.moderation.NotificationsExtension
+import io.github.jamalam360.extensions.moderation.*
 import io.github.jamalam360.extensions.user.*
 import io.github.jamalam360.util.*
 
@@ -23,7 +20,8 @@ suspend fun main() {
             add(::ModuleExtension)
             add(::LoggingExtension)
             add(::ModerationExtension)
-            add(::UtilExtension)
+            add(::ModeratorUtilityExtension)
+            add(::UserUtilityExtension)
             add(::FunExtension)
             add(::TagExtension)
             add(::NotificationsExtension)